backend/requirements.txt
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.2
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2
django-cors-headers==3.13.0
django-filter==23.2
django-import-export==3.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
# djangorestframework-jwt==1.11.0
# djangorestframework-simplejwt==4.7.2
djoser==2.1.0
et-xmlfile==1.1.0
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.3.0
mypy-extensions==1.0.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.8.0
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tablib==3.5.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
xlrd==2.0.1
xlwt==1.3.0
